refactor(docs): #704 PR-D — extract _list submodule
Moves the GET /drafts handler (`drafts_list_page`) and its
listing-page surface — row shapers (`_draft_rows`,
`_draft_list_columns`, `_DOC_TYPE_BADGE`), filter-bar state machinery
(`_DOC_TYPE_CHOICES` / `_DOC_TYPE_VALUES` / `_STATUS_VALUES` /
`_SORT_CHOICES` / `_SORT_VALUES`, `_parse_date_param`,
`_parse_filters_from_request`, `_filter_querystring`,
`_has_active_filters`), the rendered `_filter_bar` form, and the
`_drafts_table_section` HTMX wrapper — into
`app/docs/routes/_list.py` (688 lines, under the 700 cap).

`__init__.py` shrinks 2038 → 1493 (-545 lines). All moved symbols
are re-exported via `__all__` so direct imports
(`from app.docs.routes import drafts_list_page`) keep working;
`list_drafts_for_org_filtered` is also re-exported so legacy
package-level patches do not silently miss.

Test patches in `TestDraftsList*` / `TestFlashMessages` /
`TestDocTypeColumn` swap from `app.docs.routes.list_drafts_for_org_filtered`
/ `list_users` to `app.docs.routes._list.<X>` per the established
post-#704 contract (patch where used, not where defined).

Patch-path test file gets the 3 new tests for `_list` mirroring the
PR-B / PR-C blocks, pinning the contract for the moved symbol.
